Refine & optimize
