+++ client/app/visualizations/sankey/Renderer.jsx
@@ -1,28 +1,20 @@
-import { find } from 'lodash';
-import React, { useState, useEffect, useCallback } from 'react';
+import React, { useState, useEffect, useMemo } from 'react';
 import resizeObserver from '@/services/resizeObserver';
 import { RendererPropTypes } from '@/visualizations';
 
-import renderSankey from './renderSankey';
-
-function isDataValid(data) {
-  // data should contain column named 'value', otherwise no reason to render anything at all
-  return data && !!find(data.columns, c => c.name === 'value');
-}
+import initSankey from './initSankey';
 
 export default function Renderer({ data }) {
   const [container, setContainer] = useState(null);
 
-  const render = useCallback(() => {
-    if (isDataValid(data)) {
-      renderSankey(container, data.rows);
-    }
-  }, [container, data]);
+  const render = useMemo(() => initSankey(data), [data]);
 
   useEffect(() => {
     if (container) {
-      render();
-      const unwatch = resizeObserver(container, render);
+      render(container);
+      const unwatch = resizeObserver(container, () => {
+        render(container);
+      });
       return unwatch;
     }
   }, [container, render]);
